restructure add, edit and remove functions
