+++ src/index.html
@@ -19,7 +19,7 @@
       </div>
       <hr>
       <div class="mb-3">
-        <input type="text" placeholder="Add to your list..." class="form-control" id="inputTask">
+        <input type="text" placeholder="Add to your list..." class="form-control add-task" id="inputTask">
       </div>
       <div id="list"></div>
       <div class="list-footer d-flex justify-content-center py-3">
